// apps/api/src/modules/branch-user/domain/enums/permission.enum.ts
export enum Permission {
  BRANCH_USER_CREATE = 'BRANCH_USER_CREATE',
  BRANCH_USER_READ = 'BRANCH_USER_READ',
  BRANCH_USER_UPDATE = 'BRANCH_USER_UPDATE',
  BRANCH_USER_DELETE = 'BRANCH_USER_DELETE',
  BRANCH_USER_ACTIVATE = 'BRANCH_USER_ACTIVATE',
  BRANCH_USER_DEACTIVATE = 'BRANCH_USER_DEACTIVATE',
  BRANCH_USER_ASSIGN_ROLE = 'BRANCH_USER_ASSIGN_ROLE',
  BRANCH_USER_ASSIGN_PERMISSION = 'BRANCH_USER_ASSIGN_PERMISSION',
  STUDENT_CREATE = 'STUDENT_CREATE',
  STUDENT_READ = 'STUDENT_READ',
  STUDENT_UPDATE = 'STUDENT_UPDATE',
  STUDENT_DELETE = 'STUDENT_DELETE',
  COURSE_CREATE = 'COURSE_CREATE',
  COURSE_READ = 'COURSE_READ',
  COURSE_UPDATE = 'COURSE_UPDATE',
  COURSE_DELETE = 'COURSE_DELETE',
  PAYMENT_CREATE = 'PAYMENT_CREATE',
  PAYMENT_READ = 'PAYMENT_READ',
  PAYMENT_UPDATE = 'PAYMENT_UPDATE',
  PAYMENT_DELETE = 'PAYMENT_DELETE',
  REPORT_READ = 'REPORT_READ',
}